reorganize capture module
