+++ examples/capture_send.rs
@@ -1,8 +1,8 @@
 extern crate timely;
 
 use std::net::TcpStream;
-use timely::dataflow::operators::{Capture, ToStream};
-use timely::dataflow::operators::capture::EventWriter;
+use timely::dataflow::operators::ToStream;
+use timely::dataflow::operators::capture::{Capture, EventWriter};
 
 fn main() {
     timely::execute_from_args(std::env::args(), |worker| {
